Proect AI/requirements.txt
ccxt
numpy
pandas
pyyaml
websockets
aiohttp
asyncio
python-telegram-bot
fastapi
uvicorn
scikit-learn
torch
sqlalchemy
matplotlib
dash
